+++ website/tests/mocks/styleMock.js
@@ -1,2 +1,0 @@
-// Mock CSS imports for tests
-module.exports = {};